+++ examples/sphere_diameter.jl
@@ -1,4 +1,8 @@
 #! /usr/bin/env julia
+
+using Pkg
+Pkg.activate(@__DIR__)
+
 #
 # examples/sphere_diameter.jl
 #